create settings.py
