+++ log/log_config.py
@@ -1,5 +1,5 @@
 from .log import configure_logging
-
+from functools import wraps
 import logging
 import structlog
 from logging.handlers import TimedRotatingFileHandler
@@ -9,7 +9,9 @@ def proc(logger, method_name, event_dict):
     # print("I got called with", event_dict)
     return repr(event_dict)
 
+
 configure_logging(proc)
+
 
 def log_config(logger_name, filename):
     logger = structlog.get_logger(logger_name)
@@ -18,3 +20,10 @@ def log_config(logger_name, filename):
     logger.addHandler(file_handler)
     # logger.info('Тестовый запуск логгирования')
     return logger
+
+
+def log_default(func):
+    @wraps(func)
+    def call(*args, **kwargs):
+        return func(*args, **kwargs)
+    return call
