+++ server/server.py
@@ -4,17 +4,9 @@ import structlog
 
 from client.client import Client
 from log.log_config import log_config
-
-ENCODING = 'utf-8'
-MAX_MSG_SIZE = 640
+from settings import send_json
 
 logger = log_config('server', 'server.log')
-
-
-def send_json(func):
-    def inner(*args, **kwargs):
-        return json.dumps(func(*args, **kwargs)).encode(ENCODING)
-    return inner
 
 
 class Server:
